backend/requirements.txt
Django==5.*
django-ninja==1.*
psycopg2-binary==2.9.11
pydantic-settings==2.*
django-storages
boto3
redis
django-redis
django-celery-results
celery
langchain-qdrant
langchain-google-genai
qdrant-client